misc: rename to github-sponsors.ts
